requirements-prod.txt
# Production dependencies - pinned versions for fast resolution and reproducible builds

fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic>=2.0.0
supabase==2.10.0
reductoai>=0.16.0
